requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.16
django-ckeditor==6.0.0
django-cors-headers==3.10.1
django-filter==21.1
django-extra-fields
django-js-asset==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.0.0
django-cleanup==6.0.0
drf-yasg==1.20.0
ecdsa==0.17.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow
psycopg2-binary==2.8
python-dateutil==2.8.1
pyasn1==0.4.8
PyJWT==1.7.1
pymongo==3.12.1
pyparsing==3.0.9
python-jose==3.3.0
pytz==2022.1
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==1.2.0
django-taggit==2.0.0
boto3
django-modeltranslation==0.13